requirements.txt
requests
lxml
spacy
scispacy
sentence-transformers
numpy<2.0
biopython<=1.79
networkx<3.0
protobuf<5.0

